requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
pandas
streamlit
python-dotenv


